+++ MammothSunflower.java
@@ -1,9 +1,8 @@
 /* Pat Eizenga
  * 2024-09-07
  * Description: MammothSunflower class. The first seed you start with. WIP
- * 
  */
-package sunflowerSimulator;
+package typesOfFlowers;
 
 /**
  * The MammothSunflower class will create MammothSunflower objects, each holding different stats
@@ -11,72 +10,23 @@ package sunflowerSimulator;
  * A variety of stats will be needed for each MammothSunflower. Accessed through journal, check, 
  * or backpack.
  */
-public class MammothSunflower {
-	/**The growthStage of the MammothSunflower, starting with "Seed"*/
-	private String growthStage;
-	/**The daysPlanted of this particular MammothSunflower*/
-	private int daysPlanted;
-	/**The durability of this particular MammothSunflower, decrements until 0 and then destroys this MammothSunflower*/
-	private double durability;
-	/**Players can choose to eat sunflower seeds (only sunflowers, not other seed types and only while in "Seed" form)*/
+public class MammothSunflower extends Flower {
 	private int NRGRestored;
-	/**Cost of a MammothSunflower, modified by form*/
-	private double cost;
-	
-	
 	
 	/**Constructs a new MammothSunflower object
+	 * @param name of the sunflower "Mammoth Sunflower for now WIP 7/27/2024
 	 * @param growthStage "Seed", "Seedling", "Bloomed", "Matured", or "Withered"
 	 * @param daysPlanted increments one day at a time, initializes at 1
 	 * @param durability decrements with negative events, very sturdy flower
 	 * @param nRGRestored only when in "Seed" form
 	 * @param cost when bought or sold in various forms
 	 */
-	public MammothSunflower(String growthStage, int daysPlanted, double durability, int NRGRestored, double cost) {
-		setGrowthStage(growthStage);
-		setDaysPlanted(daysPlanted);
-		setDurability(durability);
+	public MammothSunflower(String name, String growthStage, int daysPlanted, double durability, int NRGRestored, double cost) {
+		super(name, growthStage, daysPlanted, durability, cost);
 		setNRGRestored(NRGRestored);
-		setCost(cost);
 	}
 
-	/**Returns the growthStage of this particular MammothSunflower
-	 * @return the growthStage when called
-	 */
-	public String getGrowthStage() {
-		return growthStage;
-	}
-	
-	/**Sets growthStage at certain intervals. Easiest flower.
-	 * @param growthStage the growthStage to set
-	 */
-	public void setGrowthStage(String growthStage) {
-		this.growthStage = growthStage;
-	}
-	/**Returns the daysPlanted of this particular MammothSunflower
-	 * @return the daysPlanted when checked in garden plot
-	 */
-	public int getDaysPlanted() {
-		return daysPlanted;
-	}
-	/**Sets the daysPlanted of this particular MammothSunflower
-	 * @param daysPlanted the daysPlanted to set increments when going to bed
-	 */
-	public void setDaysPlanted(int daysPlanted) {
-		this.daysPlanted = daysPlanted;
-	}
-	/**Returns the current durability of this particular MammothSunflower
-	 * @return the durability of this sunflower, decrements until 0 is reached
-	 */
-	public double getDurability() {
-		return durability;
-	}
-	/**Sets the durability of this particular MammothSunflower
-	 * @param durability the durability to set decrements upon negative events
-	 */
-	public void setDurability(double durability) {
-		this.durability = durability;
-	}
+
 	/**Returns an NRG amount restored for Player1 when eaten in Seed form
 	 * @return the nRGRestored with input validation
 	 */
@@ -92,18 +42,9 @@ public class MammothSunflower {
 	/**Returns the cost of the sunflower in different forms. (Seed, seedling, bloomed, matured, withered. And buy or sell)
 	 * @return the cost of this particular MammothSunflower
 	 */
-	public double getCost() {
-		return cost;
-	}
-	/**Sets the cost of the sunflower depending on form/
-	 * @param cost the cost to set can also be influenced by items/gardening level
-	 */
-	public void setCost(double cost) {
-		this.cost = cost;
-	}
-	
+
 	/**Returns true if the seed is planted.
-	 * @return true if seed is planted, return false if not planted
+	 * @return true if seed is planted, return false if not planted WIP 7/27/2024
 	 */
 	public boolean isPlanted() {
 		if (getDaysPlanted() > 0)
@@ -111,7 +52,7 @@ public class MammothSunflower {
 			return true;
 		}
 		else 
-			return false;
+			return false;	
 	}
 	
 	/**
@@ -120,9 +61,7 @@ public class MammothSunflower {
 	 */
 	@Override
 	public String toString() {
-		return "\nMammoth Sunflower " + getGrowthStage() + ":\nDays Planted: " + getDaysPlanted()
-				+ "\nDurability: " + getDurability() + "\nNRG Restored: " + getNRGRestored() +
-				"\nCost: " + getCost() + "\n";
+		return super.toString() + "NRG Restored if eaten: " + getNRGRestored() + "\n"; 
 	}
 	
 }
